Move deadline budgets into core
Constraint: Session initialization and future tool cancellation share a deadline budget primitive.
Rejected: Keep budget in src/utils | the primitive has a core lifecycle role and should not extend the utility bucket.
Confidence: high
Scope-risk: moderate
Directive: Import shared deadline budgets from src/core/deadline/budget.ts.
Tested: npm test -- --runTestsByPath tests/core/deadline/budget.test.ts tests/cdp/session-init-budget.test.ts tests/cdp/active-probe.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check
Not-tested: GitHub PR checks not yet run
